# WebAPILabs/API/Courses.API/Models/Student.cs
namespace Courses.API.Models
{
    public class Student
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public int Age { get; set; }
        public List<Course> Courses { get; set; } = new();
    }
}